% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbdata.R
\name{fbdata_results}
\alias{fbdata_results}
\title{Fetch football match results from football-data.co.uk}
\usage{
fbdata_results(league = "E0", season = "2425")
}
\arguments{
\item{league}{League code: "E0" (EPL), "E1" (Championship), "E2" (League 1),
"E3" (League 2), "SP1" (La Liga), "D1" (Bundesliga), "I1" (Serie A),
"F1" (Ligue 1), "N1" (Eredivisie), "P1" (Liga Portugal),
"SC0" (Scottish Premiership), "B1" (Jupiler League)}

\item{season}{Two-digit season code: "2324" = 2023/24, "2425" = 2024/25.
Format is last two digits of start and end year.}
}
\value{
tibble: div, date, home_team, away_team, fthg (full-time home goals),
ftag (full-time away goals), ftr (FT result H/D/A), hthg, htag, htr,
referee, hs (home shots), as_ (away shots), hst, ast, hf, af, hc, ac,
hy, ay, hr, ar
}
\description{
Downloads CSV results for a given league and season.
}
